dia1/requirements.txt
pyjwt
flask-jwt-extended